Move MVC DTOs to separate folder
