+++ BookHiveDB.Domain/BookHiveDB.Domain.csproj
@@ -8,4 +8,8 @@
     <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="7.0.9" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Dtos\Rest\BookShop\" />
+  </ItemGroup>
+
 </Project>
