+++ BookHiveDB.Domain/Dtos/Mvc/BookClubMembersDTO.cs
@@ -1,8 +1,8 @@
-﻿using BookHiveDB.Domain.DomainModels;
+﻿using System.Collections.Generic;
+using BookHiveDB.Domain.DomainModels;
 using BookHiveDB.Domain.Identity;
-using System.Collections.Generic;
 
-namespace BookHiveDB.Domain.DTO
+namespace BookHiveDB.Domain.Dtos.Mvc
 {
     public class BookClubMembersDTO
     {
