+++ BookHiveDB.Domain/Dtos/Mvc/TopicPosts.cs
@@ -1,11 +1,8 @@
-﻿using BookHiveDB.Domain.DomainModels;
+﻿using System.Collections.Generic;
+using BookHiveDB.Domain.DomainModels;
 using BookHiveDB.Domain.Identity;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
-namespace BookHiveDB.Domain.DTO
+namespace BookHiveDB.Domain.Dtos.Mvc
 {
     public class TopicPosts
     {
