+++ BookHiveDB.Domain/Dtos/Rest/Book/BookDto.cs
