+++ BookHiveDB.Domain/Dtos/Rest/Book/CreateBookDto.cs
