+++ BookHiveDB.Domain/Dtos/Rest/BookInShoppingCartDTO.cs
