+++ BookHiveDB.Domain/Dtos/Rest/BookShopDTO.cs
