+++ BookHiveDB.Service/Implementation/BookService.cs
@@ -6,6 +6,7 @@ using BookHiveDB.Service.Interface;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using BookHiveDB.Domain.Dtos.Mvc;
 using BookHiveDB.Domain.Enumerations;
 using BookHiveDB.Domain.Models;
