+++ BookHiveDB.Service/Implementation/ShoppingCartService.cs
@@ -7,6 +7,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using BookHiveDB.Domain.Dtos.Mvc;
 
 namespace BookHiveDB.Service.Implementation
 {
