+++ BookHiveDB.Service/Interface/IShoppingCartService.cs
@@ -1,5 +1,6 @@
 ﻿using BookHiveDB.Domain.DTO;
 using System;
+using BookHiveDB.Domain.Dtos.Mvc;
 
 namespace BookHiveDB.Service.Interface
 {
