+++ BookHiveDB.Web/Controllers/Mvc/BookClubController.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using BookHiveDB.Domain.DomainModels;
 using BookHiveDB.Domain.DTO;
+using BookHiveDB.Domain.Dtos.Mvc;
 using BookHiveDB.Domain.Identity;
 using BookHiveDB.Service.Interface;
 using Microsoft.AspNetCore.Identity;
