+++ BookHiveDB.Web/Controllers/Mvc/TopicController.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Threading.Tasks;
 using BookHiveDB.Domain.DTO;
+using BookHiveDB.Domain.Dtos.Mvc;
 using BookHiveDB.Domain.Identity;
 using BookHiveDB.Service.Interface;
 using Microsoft.AspNetCore.Identity;
