+++ BookHiveDB.Web/Views/Account/Register.cshtml
@@ -1,5 +1,5 @@
 ﻿@using Microsoft.AspNetCore.Mvc.TagHelpers
-@model BookHiveDB.Domain.DTO.UserRegistrationDto
+@model BookHiveDB.Domain.Dtos.Mvc.UserRegistrationDto
 @{
     ViewData["Title"] = "Register";
     Layout = null; ;
