+++ BookHiveDB.Web/Views/BookClub/Details.cshtml
@@ -1,5 +1,5 @@
 ﻿@using Microsoft.AspNetCore.Mvc.TagHelpers
-@model BookHiveDB.Domain.DTO.BookClubWithTopics
+@model BookHiveDB.Domain.Dtos.Mvc.BookClubWithTopics
 
 @{
     ViewData["Title"] = "Details";
