+++ BookHiveDB.Web/Views/BookClub/GetBookClubMembersPage.cshtml
@@ -1,5 +1,5 @@
 ﻿@using Microsoft.AspNetCore.Mvc.TagHelpers
-@model BookHiveDB.Domain.DTO.BookClubMembersDTO
+@model BookHiveDB.Domain.Dtos.Mvc.BookClubMembersDTO
 
 @{
     ViewData["Title"] = "Members";
