+++ BookHiveDB.Web/Views/BookClub/GetBookclubMembershipRequestsPage.cshtml
@@ -1,5 +1,5 @@
 ﻿@using Microsoft.AspNetCore.Mvc.TagHelpers
-@model BookHiveDB.Domain.DTO.BookClubInvitations
+@model BookHiveDB.Domain.Dtos.Mvc.BookClubInvitations
 
 @{
     ViewData["Title"] = "Member Requests";
