+++ BookHiveDB.Web/Views/BookClub/Index.cshtml
@@ -1,5 +1,5 @@
 ﻿@using Microsoft.AspNetCore.Mvc.TagHelpers
-@model BookHiveDB.Domain.DTO.BookClubsLoggedInUser
+@model BookHiveDB.Domain.Dtos.Mvc.BookClubsLoggedInUser
 
 @{
     ViewData["Title"] = "Index";
