+++ BookHiveDB.Web/Views/Books/AddBookToCart.cshtml
@@ -1,5 +1,5 @@
 ﻿@using Microsoft.AspNetCore.Mvc.TagHelpers
-@model BookHiveDB.Domain.DTO.AddToShoppingCartDto
+@model BookHiveDB.Domain.Dtos.Mvc.AddToShoppingCartDto
 
 @{ ViewData["Title"] = "Index";
                 Layout = "~/Views/Shared/_Layout.cshtml"; }
