+++ BookHiveDB.Web/Views/Home/Index.cshtml
@@ -1,5 +1,5 @@
 ﻿@using Microsoft.AspNetCore.Mvc.TagHelpers
-@model BookHiveDB.Domain.DTO.UserLoginDto
+@model BookHiveDB.Domain.Dtos.Mvc.UserLoginDto
 @{
     ViewData["Title"] = "Home Page";
     Layout = null;
