+++ BookHiveDB.Web/Views/ShoppingCart/Index.cshtml
@@ -1,7 +1,7 @@
 ﻿@using Microsoft.Extensions.Options
 @using Microsoft.AspNetCore.Mvc.TagHelpers
 @inject IOptions<BookHiveDB.Domain.StripeSettings> Stripe
-@model BookHiveDB.Domain.DTO.ShoppingCartDto
+@model BookHiveDB.Domain.Dtos.Mvc.ShoppingCartDto
 
 
 <div class="container">
