+++ BookHiveDB.Web/Views/Topic/GetTopicById.cshtml
@@ -1,5 +1,5 @@
 ﻿@using Microsoft.AspNetCore.Mvc.TagHelpers
-@model BookHiveDB.Domain.DTO.TopicPosts
+@model BookHiveDB.Domain.Dtos.Mvc.TopicPosts
 
 @{
     ViewData["Title"] = "Topic";
