+++ BookHiveDB.Web/Views/User/GetMyBookClubs.cshtml
@@ -1,5 +1,5 @@
 ﻿@using Microsoft.AspNetCore.Mvc.TagHelpers
-@model BookHiveDB.Domain.DTO.BookClubsLoggedInUser
+@model BookHiveDB.Domain.Dtos.Mvc.BookClubsLoggedInUser
 @{
     ViewData["Title"] = "My Bookclubs";
 }
